Extract function `check_incompatible_packages`
